Move `Negate` proptest to new `negate` module
